quickstart-vdk: ignore explore page and widgets in frontend
Why?

The explore page and widgets are not relevant to users outside of VMWare
This also fixes the bug where the frontend does not load on running
quickstart-vdk due to missing configuration

What?

Ignore the explore page and widgets in the control service helm chart

How has this been tested?

Configuration tested locally
Helm chart tested in CI

What type of change are you making?

Bug fix (non-breaking change which fixes an issue) or a cosmetic change/minor improvement

Signed-off-by: Dilyan Marinov <mdilyan@vmware.com>