control-service: add ability to check if docker image exists in ecr (#1977)
what: added functionality in the DockerRegistryService class which
checks if the data job image exists in a ECR style repository. This
functionality was previously unimplemented. Adopted the latest version
of the job builder - 1.3.3, which can handle creating images with
temporary amazon credentials too to be used in the integration tests.

why: as a control-service user needed to migrate to a different
repository we noticed that this functionality was missing, had it been
present the migration could have been quicker.

testing: modified existing integration test to utilize the new method
while checking if image is created in the database. New method is also
covered by the modified integration test since, it's called within the
build image method.

---------

Signed-off-by: mrMoZ1 <mzhivkov@vmware.com>
Co-authored-by: github-actions <>