vdk-dag: DAGs propagate their execution type to their component jobs (#2080)
What:
Currently, jobs ran by DAGs are labelled manual executions. This is due
to the execution being triggered through the Execution API. Ideally, we
would like that if a DAGs is a manual execution, then all component jobs
are also manual execution, and likewise for scheduled runs. This change
achieves this.

Why: To achieve consistency with the data jobs' execution type.

Testing Done: all tests are passing, introduced unit test

Signed-off-by: Yoan Salambashev <ysalambashev@vmware.com>