control-service: Only CLI executions are "Manual" (#1763)
Currently, all executions started through the Execution API are labelled
as "Manual". This makes it confusing in cases where users use the
Execution
API for starting executions not through the CLI, for example through the
Airflow integration or by a Meta Job - users will still see the
execution
as "Manual" in the UI.

This change ensures that only executions which were ran through the
VDK CLI are labelled as "Manual". Further work is required for
extracting and globalizing the "vdk-control-cli" string as it
currently appears as a magic string in both the vdk-control-cli
and the control-service.

Testing done: ran data job through meta job and verified its label

Signed-off-by: Gabriel Georgiev <gageorgiev@vmware.com>

---------

Signed-off-by: Gabriel Georgiev <gageorgiev@vmware.com>
Co-authored-by: github-actions <>
Co-authored-by: murphp15 <murphp15@tcd.ie>