docs: add getting started section for quickstart-vdk and ui (#2019)
## What

Add getting started section for quickstart-vdk and ui. Section gives an
overview on how to get started with just vdk-core and quickstart-vdk.
It's meant as a brief overview and links to more in-depth guides, e.g.
wiki and videos

## Why

README.md is the front page of the vdk project and it lacks a good
getting started section. The current getting started section just links
to the wiki.

## How was this tested

n/a

## What kind of change is this

docs

Co-authored-by: Dilyan Marinov <mdilyan@vmware.com>