frontend: ping frontend on docker image release
Why

The frontend is now part of quickstart-vdk.
On changes to the docker image, the helm
chart needs to be redeployed. We need to
make sure that the frontend is operational
once the helm chart is deployed.

What

Trigger deployment of test control service on frontend changes
merged into main
Ping the frontend using heartbeat once control service is deployed
Tag the frontend image as stable once heartbeat test passes

How has this been tested

CI

What kind of change is this

feature/non-breaking

Signed-off-by: Dilyan Marinov <mdilyan@vmware.com>