control-service: use git for images (#2097)
# Why

The most common reason for integration test failures is them being
unable to pull the image from docker hub because of throttling limits.

In this PR I pull them from github instead which doesn't have the same
restrictions.

**The newest versions of images aren't automatically published to github
and so we need to complete the following ticket asap to ensure the test
coverage is always testing the latest version of the product.**
https://github.com/vmware/versatile-data-kit/issues/2102

# What
Manually uploaded images to github. 
Updated integration test properties. 

# How was this tested
Locally and CICD

---------

Signed-off-by: murphp15 <murphp15@tcd.ie>