control-serivice: disable authorization  on deploy-testing-pipelines-service
At present, we employ an authorization webhook in our CICD repository, directing to https://httpbin.org/post, which is designed to consistently return a 200 status. However, we've observed a high degree of instability in this arrangement, as documented in https://github.com/vmware/versatile-data-kit/issues/2081.

While it's advantageous to exercise this pathway during the VDK heartbeat process, it isn't essential. This coverage is sufficiently provided through our integration and unit tests, thereby reducing the necessity of the authorization webhook's stability.