control-service: increase CICD deployment resources
This change increases the CPU and memory resource limits and requests
for the test deployment of the Control Service .

The previous configuration with CPU and memory limits of 500m and 500Mi
respectively, and equal requests, has been changed. The new
configuration sets the CPU limit to 2000m (2 cores), the memory limit to
2000Mi (approximately 2 GB), the CPU request to 1000m (1 core), and the
memory request to 1000Mi (approximately 1 GB).

This change aims to improve the performance and stability of the test
environment

Signed-off-by: Antoni Ivanov <aivanov@vmware.com>