control-service: if a test is in a bad state if fails straight away (#2098)
# Why
If a test goes into a final state that doesn't match the state we are
looking for. we can spend a couple of minutes waiting and checking does
it go to the state we want.

examples of this. 
We want a job to end with platform error. 
The job reports user error. 
we will spend 5 minutes checking is it user error. 

Total waste of time. 

# What
I have added in a fail fast clause that says if the job is not in a
running state and not in the final state we want then fail fast.



# How was this tested
Locally

---------

Signed-off-by: murphp15 <murphp15@tcd.ie>
Co-authored-by: github-actions <>