vdk-control-cli, vdk-jupyter: Allow for creating a job with a notebook step
Currently, when creating a job through the jupyter UI, the created job
has a sql step and a python step. We'd like a job created that way to
have only a notebook inside, as steps will be contained within that notebook.

This required changes to both the control cli module, as well as the jupyter
python module. An alternative way was possible which left the control cli
untounched, however that path was not chosen since it would require patching
module namespaces during runtime, which is prone to bugs and is to be avoided
if possible.

Testing done: tested locally by creating a new job through jupyter UI and observing
the notebook step inside

Signed-off-by: Gabriel Georgiev <gageorgiev@vmware.com>