Why:
Add functionality to the insert processing templates in order to allow
quality checks to be made before the data is inserted into the target
table.

Currently, the checks done on the insert template processing step are
not covering if the semantics of the data is correct. Therefore, bad
data could go into the target table which could be unwanted behavior.

More details explained in
https://github.com/vmware/versatile-data-kit/issues/1361

What:
-Adding functionality to handle the insert template behavior if the
user provides checks

Tests: provided positive and negative regression tests

Signed-off-by: Stefan Buldeev sbuldeev@vmware.com