vdk-core: Update JobConfig to match vdk-control-cli JobConfig
Currently, there is code duplication between the vdk-core's JobConfig and vdk-control-cli's
JobConfig. This is not sustainable, as when a change in one component is made, this change
would need to be ported to the other component.

This change updates the vdk-core's JobConfig, as vdk-core is the main component of the
Versatile Data Kit SDK. In a separate PR, the JobConfig of vdk-control-cli will be removed
to remove the code duplication.

Testing Done: CI/CD

Signed-off-by: Andon Andonov <andonova@vmware.com>