vdk-control-cli: Allow extensions to specify a sample job (#2177)
The purpose of this change is the following: some extensions to VDK
might use vdk-control-cli as a dependency to create new jobs of a
certain type through the JobCreate class. Currently, there is a way to
specify what job is created, which is the
VDK_CONTROL_SAMPLE_JOB_DIRECTORY, however this config variable is aimed
at users which want to create a job template which is specific to their
environment, and must be in the form of a path in their file system. An
extension cannot use this to create a new sample job. This change allows
for this by extending the JobCreate class' create method with a new
parameter which can be passed when the method is called. This parameter
must be a path to a job directory and must contain the necessary job files
which will be created when running creating a job.

Testing done: unit test

---------

Signed-off-by: Gabriel Georgiev <gageorgiev@vmware.com>