control-service: secrets service implementation
Iterative change for VEP-1493.

The service is still not publich/marked as not implemented due to
lack of tests which are going to be implmented as a next step.

Added the ability to optionallly configure a hashicorp vault instance,
through a feature flag, uri token settings.

The change uses the spring cloud library as a client library for
connecting to Hashicorp Vault.

I've opted out of using a Vault based Spring CRUD repositroy and
other Spring Vault facilities as the 2.x version of the spring
libraries cannot work with a vault k/v secrets storage v.2 which is
now the default and migrating to spring 3.0 would require a lot of
other changes across the whole service.

Signed-off-by: Dako Dakov <ddakov@vmware.com>