[Bug fix] vdk-core: Fix _get_contacts in JobConfig
When the logic from vdk-control-cli's JobConfig class was synced with the
vdk-core's JobConfig, the `._get_contacts()` method logic was not properly
ported, and this caused issues with config.ini files, where email addresses
are not separated with `;`.

This change fixes the introduced bug. Additionally, the job config tests from
the vdk-control-cli are  refactored and moved to the vdk-core.

Testing Done: New unit tests.

Signed-off-by: Andon Andonov <andonova@vmware.com>