control-service: make kubernetes service easy  to test. (#2249)
# What
The KubernetesService is by far the biggest class in control-plane. 

It has much too much responsibilities. 

It is responsible for initialing the class that it will use. 
For example it is responsible for creating the k8s client. 
This is against the spring model where each component should be setup
with all the components it needs.
But also it makes it hard to test, because we can't pass in mock
clients.

In this PR I create the client outside of the kuberentesService and
autowire it into the service.
I also do this for the batch clients. 


In a previous
PR(https://github.com/vmware/versatile-data-kit/pull/2207#discussion_r1221372123)
it was pointed out that there is a bug in the kubernetesService where
500s from the cluster are misclassified as the job not existing.
To write a test for that bug is not easy because we are creating the
clients within the class.
After this PR is merged it will be much easier to test that bug. 




# How
I have taken the approach that more properties in kuberentes service
should be final.

In the PR it should be obvious that this change is better for tests and
production code.

In the old code we were calling initBatchV1Client() before every call to
kuberentes just to create the client again which was totally needless
because it should have been a singleton.

Also in the tests we were often creating a mock to mock a single
function to return another mock. All this noise is gone from the tests.



# How how this tested? 
Local and CICD

---------

Signed-off-by: murphp15 <murphp15@tcd.ie>
Co-authored-by: github-actions <>