vdk-core: Add python_version configuration to config-help
As part of the initiative to support multiple python versions for data job
deployments, a new configuration property, `python_version` was introduced.
This configuration, however, was not initially added to the output of the
`vdk config-help` command.

This change adds the `python_version` configuration to the JobConfigIniPlugin,
as it is related to the config.ini file setup.

Testing Done:
Tested locally by running `vdk config-help` and verifying that the configuration
is present in the output.

Signed-off-by: Andon Andonov <andonova@vmware.com>