control-service: remove spammy logs
The Control Service has some pretty spammy logs that are making it
harder to pick the important things.

For example

```
?[30m2023-06-19 06:38:27,637?[0;39m ?[39mDEBUG?[0;39m [OpId:]
?[34mthread-3  ?[0;39m ?[33mc.v.t.s.locks.CustomLockProvider?[0;39m
Releasing lock watchJobs_schedulerLock
?[30m2023-06-19 06:38:27,637?[0;39m ?[39mDEBUG?[0;39m [OpId:]
?[34mthread-3  ?[0;39m ?[33mc.v.t.s.locks.CustomLockProvider?[0;39m Lock
watchJobs_schedulerLock is released
```

Those are not really needed in most cases . They still might be useful
if troubleshooting some issue with the locks so I am disabling them in
applicaiton.properties configuraiton (which can be overriden if needed)

```
?[30m2023-06-19 06:38:27,634?[0;39m ?[34mINFO ?[0;39m [OpId:]
?[34mthread-3  ?[0;39m ?[33m.t.s.k.DataJobsKubernetesService?[0;39m
Start watching jobs with labels: {com.vmware.taurus/type=DataJob}
```

Move the information about hte log when DataJobMonitor is initialized.
Not needed every time.

```
?[30m2023-06-19 06:38:27,637?[0;39m ?[34mINFO ?[0;39m [OpId:]
?[34mthread-3  ?[0;39m ?[33m.t.s.k.DataJobsKubernetesService?[0;39m
Failed to list jobs for watching. Error was:
```

This is spammy but for some reason the error is not printed. So I tried
to use KubernetseExceptoin which provides more details.
Also `e.toString()` is prefered over `e.getMessage()` since the first
one prints at least the exception class (even if there's no message)

Signed-off-by: Antoni Ivanov <aivanov@vmware.com>