frontend: use @versatiledatakit/shared rc-builds for data-pipelines
Why?

Linking shared-components to data-pipelines using npm fails to produce
a working UI docker image

What?

Introduce rc-versions for @versatiledatakit/shared
On a successful build of shared, and rc-version is pushed to npm registry
The data-pipelines build pulls the latest rc-version when running, instead of
linking to the local installation of shared

Versioning schema follows semver

<major>.<minor>.<patch>-rc.<rc-number>

The rc-number is incremented based on the latest rc-number for the pipeline-id.
This helps avoid conflicts between different pipelines running builds for shared

Regular shared versions are tagged with latest in npm registry
rc-versions are tagged with rc

How was this tested?

Ran a build in CI and tested that the docker image starts locally
as part of quickstart-vdk by changing the deployment tag

https://gitlab.com/vmware-analytics/versatile-data-kit/-/pipelines/909682172

What kind of change is this?

Feature/non-breaking

Signed-off-by: Dilyan Marinov <mdilyan@vmware.com>