control-service: remove needless variable from gitlab variables (#2334)
# Why
It makes everything simpler. 
We are loading the config from a gitlab variable when the same content
can be loaded from a file in the repo.

There is not reason to use a gitlab variable and it makes us less
dependant on gitlab.


After this is merged I will remove the variable in gitlab. 


# How was this tested
NA

Signed-off-by: murphp15 <murphp15@tcd.ie>
Co-authored-by: paulm2 <paulm2@vmware.com>