ci: slack notification on pipeline failure
Why?

In order to set up CI on-call for vdk,
visible and reliable notifications are required.
The notifications should alert on pipeline failure
and provide the necessary information for troubleshooting.

What?

Call an incoming webhook for a custom slack application on
pipeline failure. This sends a message to the corresponding channel.

How was this tested?

Ran in pipeline and sent notification to debug channel.
Note that the env variable for the notification hook will have to be changed
after this is merged.

https://gitlab.com/vmware-analytics/versatile-data-kit/-/pipelines/914912201

What kind of change is this

Feature/non-breaking

Signed-off-by: Dilyan Marinov <mdilyan@vmware.com>