vdk-jupyter: Create init cell when opening new notebook
This PR ensures that the necessary init cell is created
automatically when opening an empty notebook. The specific
logic goes as follows: first, we check if this is a new
notebook by checking that it only has 1 empty cell. Then,
we create the necessary cell, insert it and delete the previous
one.

Testing done: manual

Signed-off-by: Gabriel Georgiev <gageorgiev@vmware.com>