frontend: quickstart-vdk operability tests using cypress
Why?

We test that the docker image is starting correctly,
but we don't actually test that the frontend is operational.
This has resulted in docker images where the docker image
starts, but the frontend is not operational with no way
of detecting the problem.

What?

- Move regular e2e tests into their own directory
- Deployment to test environment now depends on e2e tests running
successfully beforehand
- Introduce quickstart-smoke script
- Create basic cypress smoke test for the frontend
- Run smoke test after deployment to test environment

How was this tested?

Ran quickstart-vdk locally and ran cypress against frontend
CI

What kind of change is this?

Feature/non-breaking

Signed-off-by: Dilyan Marinov <mdilyan@vmware.com>