frontend: Fix navigation in Data Jobs
Changed navigation in Data Jobs list from navigateToUrl
to locationToURL, because telemetry events are sent only
when Angular router navigation occurs. We don't want telemetry
to be send when filters are applied in the URL as query string.
Fix for Browser pop states while User is in Data Jobs list and
going backward/forward and state from URL is propagated to the
component.
Fix for some e2e tests.

Signed-off-by: Goran Kokinovski <gkokinovski@vmware.com>