vdk-impala: fix pydantic to 1.x
A week ago pydantic released 2.0 version which is a breaking version .

https://docs.pydantic.dev/dev-v2/migration Fortunately pydantic has also
released a tool bump-pydantic which automatically upgraded the code.
That's useful though some needed to be edit manually.

Also in tests use `cli_assert(not res.exception, res)` instead of assert
to print more details about the job failure.