vdk-impala: support also pydantic 1.0
Make sure vdk-impala works with pydantic 1.0 as this is still used and
some libraries do not support the new version.

In hindsight I probably should have pinned to 1.0 only . But I did not
and it's good idea to support the new version otherwise at some point in
the future it will cause conflicts and problems as other library
adopted.