vdk-smarter: introduce simple open ai plugin (#2351)
Very simple plugin that reviews every query and scores it using OpenAI
models. The purpose is to provide easy visibility into queries and their
quality for the user. Though technically they could potentially use
chatGPT themselves or static code analysis, that is less likely and and
harder for them

A) because some queries are generated dynamically
B) Many queries could be part of a job and take (development) time to
check.
C) the prompt engineering is something that takes time as well and
tuning as well. The prompts could be tuned by community pretty well over
time.

Inside an enterprise, this enables central data team to establish best
practices around SQL with proper review model themselves as well.

This is a POC (a Pre-Alpha development status as defined in
https://martin-thoma.com/software-development-stages/#2-pre-alpha)

---------

Signed-off-by: Antoni Ivanov <aivanov@vmware.com>