vdk-core: Allow different python versions for vdk docker images (#2346)
Currently, all vdk-related docker images (mainly quickstart-vdk) are
built with python3.7, because the python base image is hard-coded into
the Dockerfile (Dockerfile-vdk-base). This causes issues if images for
different python versions are to be built.

This change adds more flexibility, by updating the Dockerfile to accept
build argument with the python image that is to be used for building the
vdk image. It also updates the shell script that builds and releases the
vdk images to read a new environment variable, VDK_BASE_PYTHON_IMAGE,
which indicates which python version is to be used.

Testing Done: Locally executed the script and verified that the passed
python base image propagates to the Dockerfile.

Signed-off-by: Andon Andonov <andonova@vmware.com>