Truncate table before inserting data
Why:
Currently when the insert template is used with quality checks it only appends data to the staging table without truncating it at any moment. This will most probably lead to duplicated data on the following run, because old source data will be already in the staging table before appending the new one.

More details explained in
#1361

What:
-Adding truncate statement before the data is being inserted in the staging table to ensure that no leftover data from previous runs is left.

Signed-off-by: Stefan Buldeev sbuldeev@vmware.com