vdk-impala: support also pydantic 1.0 (#2368)
Ensure vdk-impala works with pydantic 1.0 as this is still used and
libraries do not support the new version.

In hindsight, I should have pinned to 1.0 only. But I did not and it's a
good idea to support the new version otherwise at some point it will
cause conflicts and problems as other libraries adopt 2.0 of pydantic

Testing Done: installed locally pydantic 1.0 . The Gitlab CI is testing
with 2
 - `pip install pydantic==1.*`
 - ../build-plugin.sh (from vdk-impala dir) 
- verified the test exited successfully. 

I toyed with the idea of adding a test for pydanitc 1.0 but it would
require some refactoring of hte plugin framework to allow that and it's
not that urgent as the changes are fairly small.