frontend: enable stable tagging
Why?

https://github.com/vmware/versatile-data-kit/issues/2332

This issue has been resolved, so we can re-enable stable
tagging for the frontend. The change lets us use our CI/CD pipeline to deliver
updates to the frontend and quickstart-vdk with sufficient
release confidence.

What?

Enable stable tagging for the frontend docker
image. This lets us deliver updates for the
frontend as part of quickstart-vdk.

Note: The stable tag is pulled each time quickstart-vdk is restarted
and the helm chart it uses re-deployed

How was this tested?

The stable tagging job ran as part of the CI before,
we just didn't have sufficient testing around
the frontend image, so we disabled it. We use
the same script for tagging with latest, so
it should be fine.

What kind of change is this?

Feature/non-breaking

Signed-off-by: Dilyan Marinov <mdilyan@vmware.com>