frontend: fix data-pipelines build scripts
Why?

The data pipelines build script does not work correctly.
This is because of an incorrect path in the ci config and typo in the install_data_pipelines script.

What?

Fix path and typo

How was this tested?

Tested the script locally to make sure the correct execution path is hit
depending on what's passed as shared components rc version

What kind of change is this?

Bugfix

Signed-off-by: Dilyan Marinov <mdilyan@vmware.com>