vdk-heartbeat: cover requirements.txt automatic installs
Upon `vdk deploy` the Control Service is responsible for installing all
external dependecies defined in requirements.txt of the data job
so they can be used during a cloud run.

Recent outage in one of our customers which caused data jobs have not
had their requirements installed. It was trace to be due to a release of
[data job base
image](https://hub.docker.com/layers/versatiledatakit/data-job-base-python-3.11)

This was not caught by CICD of the open source VDK.

This change aim to address this gap and to cover this improtnat
functionality for correct installation of data job's python
requirements.