control-service: push to multiple registries (#2381)
To avoid issues with docker limits sometimes it may be beneficial to use
ghcr.io instead for example in our CICD or in vdk-server .

This change publishes some images to both docker.io and ghcr.io

The GitLab CI configuration is changed to include a shared library for
common CI/CD tasks.
Two new scripts have been added for parsing and pushing Docker images. 
These scripts are now being used in the build process to push images to
multiple registries in.

Testing Done: Temporarily changed the pipeline to push images on PR and
verified images are pushed to both correctly. See it in
https://gitlab.com/vmware-analytics/versatile-data-kit/-/pipelines/927292918