vdk-jupyter: publish image to pip registry (#2407)
# Why
We need to publish the image to a pip registry so others can consume it
and setup is easy and doesn't require cloning the source code

# What
Most plugins are installed using the same command. which is declared in
a common gitlab file.
For building this plugin I didn't think it made sense to reuse that code
as building jupyter is a little more complicated.
So the release is totally new code. 


# How was this tested
I published to the registry and installed locally. 

Commands: 
```
pip install -U vdk_jupyterlab_extension
jupyter lab
```
Make sure the UI has the correct panels
![Screenshot 2023-07-13 at 12 32
42](https://github.com/vmware/versatile-data-kit/assets/9319000/a2484490-f367-4608-8df0-1b64debd156d)

---------

Signed-off-by: murphp15 <murphp15@tcd.ie>
Co-authored-by: paulm2 <paulm2@vmware.com>