vdk-jupyter: Create init cell when opening new notebook (#2352)
Our integration with Jupyter requires two magic commands
plus a line of Python code to be ran before it can work. Currently,
users need to copy and paste these from previous notebooks
into new ones. It would probably be better if this was automated.

This PR ensures that the necessary init cell is created
automatically when opening an empty notebook. The specific
logic goes as follows: first, we check if this is a new
notebook by checking that it only has 1 empty cell. Then,
we create the necessary cell, insert it and delete the previous
one.

Testing done: manual

Signed-off-by: Gabriel Georgiev <gageorgiev@vmware.com>