vdk-plugins: fix build of multiple plugins (#2445)
There were overall 3 issues that needed to be fixed in the build of
those plugins:

The first one is 2.0 major release of pytest-docker
(https://pypi.org/project/pytest-docker/#history) which we use for
starting docker container in some plugin tests . This impacted us
because

1. pytest-docker change default command from `docker-compose` to `docker
compose` (no dash)
2. But the library we use (docker-compose) installs only
`docker-compose` command (with a dash)

--
The other problematic library is the installation of docker-compose
itself :).
docker-compose depends on PyYaml and PyYAML is having the following
issue: https://github.com/yaml/pyyaml/issues/724 (in short a recent
release of Cython 3.0 broke PyYAML)

As a workaround, we can pin `PyYAML==5.3.1` And also remove eager
(transitive) upgrades when installing the package with pip to reduce the
chance of accidentally upgrading it.

--
Finally removed all non- testing dependencies from requirements.txt. In
one of the plugin was failing to install because it was missing install
requirements in setup.py and was not caught during testing because
requirements.txt was installing that dependency.

------

An alternative approach to fix it might have been to switch to use
[testcontainer](https://pypi.org/project/testcontainers/) (like
vdk-trino uses) instead of pytest-docker + docker-compose. But that
would require refactoring of the tests so I opted out for now.