vdk-heartbeat: cover requirements.txt automatic installs (#2393)
Upon `vdk deploy` the Control Service is responsible for installing all
external dependecies defined in requirements.txt of the data job so they
can be used during a cloud run.

A recent outage in one of our customers caused data jobs to ignore their
requirements.txt and not install any external libraries. It was trace to
be due to a release of [data job base

image](https://hub.docker.com/layers/versatiledatakit/data-job-base-python-3.11)

CICD of the open source VDK did not catch this. 

I am extending VDK heartbeat to test for the issue so we can catch it in
the future.
 
See also https://github.com/vmware/versatile-data-kit/issues/2391