vdk-control-cli: pass the printer class in JobDeploy creation and add additional memory printer (#2477)
What:Introduced an enhanced feature, the MemoryPrinter, it conserves
text for future use, instead of instant printing. Along with this,
adapted the JobDeploy class to receive a printer object instead of an
OutputFormat enum.

Why:
The inclusion of MemoryPrinter caters to use cases where an output needs
to be stored for later representation, such as in a Jupyter UI. This
modification in the JobDeploy class enables the printer object's use
beyond JobDeploy: the printer can be initialised first, utilised within
JobDeploy, and then accessed again to retrieve the stored output. This
stops the need for accessing private fields within JobDeploy and
promotes better coding practices.

Signed-off-by: Duygu Hasan [hduygu@vmware.com](mailto:hduygu@vmware.com)

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>