vdk-ipython: use package structure for file (#2499)
# Why
The ipython file was not in a package and was causing issues for our
setup.py files. They expect all python files to be in a package.
# What 
Put it in a package
# How was this tested. 

I ran the ui-tests and they were all green. I also started juypter
locally and it was able to load

Co-authored-by: paulm2 <paulm2@vmware.com>