[vdk-plugins] vdk-impala: Handle authorization errors on refresh
Currently, if a data job executes a select query, which fails due to some metadata-related
error, the error handling logic executes refresh/invalidate metadata query to sync the
table metadata and retry the initial query. This works if the data job has write access
to the table. However, if the job has only read access, the refresh/invalidate metadata
query will fail with authorization error. This will lead to overall job execution failure,
which is not ideal, as vdk, not the job itself is trying to execute an illegal operation.

This change updates the error handling logic, so that when an authorization error on
refresh/invalidate metadata is raised, vdk will re-try the query with a backoff (if the
job does not have write access, we cannot know when the metadata for the table has been
updated, and the best we can do is sleep and re-try).

Testing Done: Added test

Signed-off-by: Andon Andonov <andonova@vmware.com>