[vdk-plugin] vdk-impala: Make error retries and backoff configurable
Currently, the retries and backoff seconds are hard-coded in the ImpalaErrorHandler, which
is not ideal, as there might be users who want to have specific values for these properties.

This change adds two new configuration variables, `IMPALA_RETRIES_ON_ERROR` and `IMPALA_ERROR_BACKOFF_SECONDS`,
which are to be used for configuring how many times a failed query is to be retried, and how big the backoff
should be.

Signed-off-by: Andon Andonov <andonova@vmware.com>