vdk-sqlite: support whitespace in column names
Why?

Trying to create columns with whitespace in the name
and inserting data in them using vdk-sqlite resulted
in error.

This was most evident in examples like csv-ingestion

https://github.com/vmware/versatile-data-kit/wiki/Ingesting-local-CSV-file-into-Database

What?

Add double quotes around column names with whitespace in them when the table is created.
Replace named placeholders in insert query with question marks
Convert payload from dicts to lists to pass it to above query
Check if payload headers match column names before it's passed

How has this been tested?

Unit tests

What type of change is this?

Bugfix

Signed-off-by: Dilyan Marinov <mdilyan@vmware.com>